# backend/WandelApp.api/WandelApp.api2/bin/Release/net5.0/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "WandelApp.api",
    "version": "1.0"
  },
  "paths": {
    "/api/Friend": {
      "get": {
        "tags": [
          "Friend"
        ],
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/GetFriendModel"
              }
            }
          },
          "404": {
            "description": "Not Found",
            "schema": {
              "$ref": "#/definitions/ProblemDetails"
            }
          }
        }
      },
      "post": {
        "tags": [
          "Friend"
        ],
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/PostFriendModel"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/GetFriendModel"
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/ProblemDetails"
            }
          }
        }
      }
    },
    "/api/Friend/{id}": {
      "get": {
        "tags": [
          "Friend"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/GetFriendModel"
            }
          },
          "404": {
            "description": "Not Found",
            "schema": {
              "$ref": "#/definitions/ProblemDetails"
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/ProblemDetails"
            }
          }
        }
      },
      "put": {
        "tags": [
          "Friend"
        ],
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "string"
          },
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/PutFriendModel"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Success"
          },
          "404": {
            "description": "Not Found",
            "schema": {
              "$ref": "#/definitions/ProblemDetails"
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/ProblemDetails"
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Friend"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "204": {
            "description": "Success"
          },
          "404": {
            "description": "Not Found",
            "schema": {
              "$ref": "#/definitions/ProblemDetails"
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/ProblemDetails"
            }
          }
        }
      }
    },
...
    "/api/Pet/{id}": {
      "get": {
        "tags": [
          "Pet"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/GetPetModel"
            }
          },
          "404": {
            "description": "Not Found",
            "schema": {
              "$ref": "#/definitions/ProblemDetails"
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/ProblemDetails"
            }
          }
        }
      },
      "put": {
        "tags": [
          "Pet"
        ],
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "string"
          },
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/PutPetModel"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Success"
          },
          "404": {
            "description": "Not Found",
            "schema": {
              "$ref": "#/definitions/ProblemDetails"
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/ProblemDetails"
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Pet"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "204": {
            "description": "Success"
          },
          "404": {
            "description": "Not Found",
            "schema": {
              "$ref": "#/definitions/ProblemDetails"
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/ProblemDetails"
            }
          }
        }
      }
    },
...
    "/api/User/{id}": {
      "get": {
        "tags": [
          "User"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/GetUserModel"
            }
          },
          "404": {
            "description": "Not Found",
            "schema": {
              "$ref": "#/definitions/ProblemDetails"
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/ProblemDetails"
            }
          }
        }
      },
      "put": {
        "tags": [
          "User"
        ],
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "string"
          },
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/PutUserModel"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Success"
          },
          "404": {
            "description": "Not Found",
            "schema": {
              "$ref": "#/definitions/ProblemDetails"
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/ProblemDetails"
            }
          }
        }
      },
      "delete": {
        "tags": [
          "User"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "204": {
            "description": "Success"
          },
          "404": {
            "description": "Not Found",
            "schema": {
              "$ref": "#/definitions/ProblemDetails"
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/ProblemDetails"
            }
          }
        }
      }
    },
...
    "/api/UserPet": {
      "get": {
        "tags": [
          "UserPet"
        ],
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/GetUserPetModel"
              }
            }
          },
          "404": {
            "description": "Not Found",
            "schema": {
              "$ref": "#/definitions/ProblemDetails"
            }
          }
        }
      },
      "post": {
        "tags": [
          "UserPet"
        ],
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/PostUserPetModel"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/GetUserPetModel"
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/ProblemDetails"
            }
          }
        }
      }
    },
    "/api/UserPet/{id}": {
      "get": {
        "tags": [
          "UserPet"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/GetUserPetModel"
            }
          },
          "404": {
            "description": "Not Found",
            "schema": {
              "$ref": "#/definitions/ProblemDetails"
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/ProblemDetails"
            }
          }
        }
      },
      "put": {
        "tags": [
          "UserPet"
        ],
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "string"
          },
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/PutUserPetModel"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Success"
          },
          "404": {
            "description": "Not Found",
            "schema": {
              "$ref": "#/definitions/ProblemDetails"
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/ProblemDetails"
            }
          }
        }
      },
      "delete": {
        "tags": [
          "UserPet"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "204": {
            "description": "Success"
          },
          "404": {
            "description": "Not Found",
            "schema": {
              "$ref": "#/definitions/ProblemDetails"
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/ProblemDetails"
            }
          }
        }
      }
    }
  },
  "definitions": {
    "GetFriendModel": {
      "required": [
        "DateCreated"
      ],
      "type": "object",
      "properties": {
        "UserId": {
          "format": "uuid",
          "type": "string"
        },
        "UserFriendId": {
          "format": "uuid",
          "type": "string"
        },
        "Id": {
          "format": "uuid",
          "type": "string"
        },
        "DateCreated": {
          "format": "date",
          "type": "string"
        }
      }
    },
    "ProblemDetails": {
      "type": "object",
      "properties": {
        "type": {
          "type": "string"
        },
        "title": {
          "type": "string"
        },
        "status": {
          "format": "int32",
          "type": "integer"
        },
        "detail": {
          "type": "string"
        },
        "instance": {
          "type": "string"
        }
      },
      "additionalProperties": { }
    },
    "PostFriendModel": {
      "type": "object",
      "properties": {
        "UserId": {
          "format": "uuid",
          "type": "string"
        },
        "UserFriendId": {
          "format": "uuid",
          "type": "string"
        }
      }
    },
    "PutFriendModel": {
      "type": "object",
      "properties": {
        "UserId": {
          "format": "uuid",
          "type": "string"
        },
        "UserFriendId": {
          "format": "uuid",
          "type": "string"
        }
      }
    },
...
    "GetPetModel": {
      "required": [
        "Experience",
        "Level",
        "Name"
      ],
      "type": "object",
      "properties": {
        "Name": {
          "type": "string"
        },
        "Experience": {
          "format": "int32",
          "type": "integer"
        },
        "Level": {
          "format": "int32",
          "type": "integer"
        },
        "Id": {
          "format": "uuid",
          "type": "string"
        },
        "UserPets": {
          "type": "array",
          "items": {
            "format": "uuid",
            "type": "string"
          }
        }
      }
    },
...
    "PutPetModel": {
      "required": [
        "Experience",
        "Level",
        "Name"
      ],
      "type": "object",
      "properties": {
        "Name": {
          "type": "string"
        },
        "Experience": {
          "format": "int32",
          "type": "integer"
        },
        "Level": {
          "format": "int32",
          "type": "integer"
        },
        "UserPets": {
          "type": "array",
          "items": {
            "format": "uuid",
            "type": "string"
          }
        }
      }
    },
    "GetUserModel": {
      "required": [
        "Age",
        "CountPetsCaptured",
        "CurrentPoints",
        "Email",
        "Password",
        "UserName"
      ],
      "type": "object",
      "properties": {
        "Email": {
          "maxLength": 45,
          "minLength": 0,
          "type": "string"
        },
        "Password": {
          "maxLength": 45,
          "minLength": 0,
          "type": "string"
        },
        "UserName": {
          "maxLength": 45,
          "minLength": 0,
          "type": "string"
        },
        "Age": {
          "format": "int32",
          "type": "integer"
        },
        "Id": {
          "format": "uuid",
          "type": "string"
        },
        "CurrentPoints": {
          "format": "int32",
          "type": "integer"
        },
        "CountPetsCaptured": {
          "format": "int32",
          "type": "integer"
        },
        "NextLevelId": {
          "format": "uuid",
          "type": "string"
        },
        "NextLevelName": {
          "type": "string"
        },
        "NextLevelPointsRequired": {
          "format": "int32",
          "type": "integer"
        },
        "CurrentLevelName": {
          "type": "string"
        },
        "UserPets": {
          "type": "array",
          "items": {
            "format": "uuid",
            "type": "string"
          }
        },
        "Friends": {
          "type": "array",
          "items": {
            "format": "uuid",
            "type": "string"
          }
        },
        "UserGames": {
          "type": "array",
          "items": {
            "format": "uuid",
            "type": "string"
          }
        }
      }
    },
...
    "PutUserModel": {
      "required": [
        "Age",
        "CountPetsCaptured",
        "CurrentPoints",
        "Email",
        "Password",
        "UserName"
      ],
      "type": "object",
      "properties": {
        "Email": {
          "maxLength": 45,
          "minLength": 0,
          "type": "string"
        },
        "Password": {
          "maxLength": 45,
          "minLength": 0,
          "type": "string"
        },
        "UserName": {
          "maxLength": 45,
          "minLength": 0,
          "type": "string"
        },
        "Age": {
          "format": "int32",
          "type": "integer"
        },
        "CurrentPoints": {
          "format": "int32",
          "type": "integer"
        },
        "CountPetsCaptured": {
          "format": "int32",
          "type": "integer"
        },
        "NextLevelId": {
          "format": "uuid",
          "type": "string"
        },
        "UserPets": {
          "type": "array",
          "items": {
            "format": "uuid",
            "type": "string"
          }
        },
        "Friends": {
          "type": "array",
          "items": {
            "format": "uuid",
            "type": "string"
          }
        },
        "UserGames": {
          "type": "array",
          "items": {
            "format": "uuid",
            "type": "string"
          }
        }
      }
    },
...
    "GetUserPetModel": {
      "type": "object",
      "properties": {
        "PetId": {
          "format": "uuid",
          "type": "string"
        },
        "UserId": {
          "format": "uuid",
          "type": "string"
        },
        "Id": {
          "format": "uuid",
          "type": "string"
        }
      }
    },
    "PostUserPetModel": {
      "type": "object",
      "properties": {
        "PetId": {
          "format": "uuid",
          "type": "string"
        },
        "UserId": {
          "format": "uuid",
          "type": "string"
        }
      }
    },
    "PutUserPetModel": {
      "type": "object",
      "properties": {
        "PetId": {
          "format": "uuid",
          "type": "string"
        },
        "UserId": {
          "format": "uuid",
          "type": "string"
        }
      }
    }
  }
}